+++ src/VBox/NetworkServices/Dhcpd/DhcpOptions.h
@@ -4,15 +4,25 @@
  */
 
 /*
- * Copyright (C) 2017-2020 Oracle Corporation
+ * Copyright (C) 2017-2022 Oracle and/or its affiliates.
  *
- * This file is part of VirtualBox Open Source Edition (OSE), as
- * available from http://www.virtualbox.org. This file is free software;
- * you can redistribute it and/or modify it under the terms of the GNU
- * General Public License (GPL) as published by the Free Software
- * Foundation, in version 2 as it comes in the "COPYING" file of the
- * VirtualBox OSE distribution. VirtualBox OSE is distributed in the
- * hope that it will be useful, but WITHOUT ANY WARRANTY of any kind.
+ * This file is part of VirtualBox base platform packages, as
+ * available from https://www.virtualbox.org.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation, in version 3 of the
+ * License.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, see <https://www.gnu.org/licenses>.
+ *
+ * SPDX-License-Identifier: GPL-3.0-only
  */
 
 #ifndef VBOX_INCLUDED_SRC_Dhcpd_DhcpOptions_h
@@ -446,7 +456,7 @@ protected:
             return -1;
 
         append(dst, m_String);
-        return m_String.length();
+        return (ssize_t)m_String.length();
     }
 
 public:
@@ -550,7 +560,7 @@ protected:
             cbValue += cbItem;
         }
 
-        return cbValue;
+        return (ssize_t)cbValue;
     }
 
 public:
@@ -705,13 +715,13 @@ protected:
     virtual ssize_t encodeValue(octets_t &dst) const
     {
         dst.insert(dst.end(), m_Data.begin(), m_Data.end());
-        return m_Data.size();
+        return (ssize_t)m_Data.size();
     }
 
     virtual int decodeValue(const octets_t &src, size_t cb)
     {
         octets_t::const_iterator beg(src.begin());
-        octets_t data(beg, beg + cb);
+        octets_t data(beg, beg + (ssize_t)cb);
         m_Data.swap(data);
 
         m_fPresent = true;
