+++ src/VBox/Runtime/common/checksum/RTSha1Digest.cpp
@@ -7,24 +7,34 @@
  */
 
 /*
- * Copyright (C) 2009-2020 Oracle Corporation
+ * Copyright (C) 2009-2022 Oracle and/or its affiliates.
  *
- * This file is part of VirtualBox Open Source Edition (OSE), as
- * available from http://www.virtualbox.org. This file is free software;
- * you can redistribute it and/or modify it under the terms of the GNU
- * General Public License (GPL) as published by the Free Software
- * Foundation, in version 2 as it comes in the "COPYING" file of the
- * VirtualBox OSE distribution. VirtualBox OSE is distributed in the
- * hope that it will be useful, but WITHOUT ANY WARRANTY of any kind.
+ * This file is part of VirtualBox base platform packages, as
+ * available from https://www.virtualbox.org.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation, in version 3 of the
+ * License.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, see <https://www.gnu.org/licenses>.
  *
  * The contents of this file may alternatively be used under the terms
  * of the Common Development and Distribution License Version 1.0
- * (CDDL) only, as it comes in the "COPYING.CDDL" file of the
- * VirtualBox OSE distribution, in which case the provisions of the
+ * (CDDL), a copy of it is provided in the "COPYING.CDDL" file included
+ * in the VirtualBox distribution, in which case the provisions of the
  * CDDL are applicable instead of those of the GPL.
  *
  * You may elect to license modified versions of this file under the
  * terms and conditions of either the GPL or the CDDL or both.
+ *
+ * SPDX-License-Identifier: GPL-3.0-only OR CDDL-1.0
  */
 
 
@@ -57,7 +67,7 @@ RTR3DECL(int) RTSha1Digest(void* pvBuf, size_t cbBuf, char **ppszDigest, PFNRTPR
     RTSha1Init(&Ctx);
 
     /* Buffer size for progress callback */
-    double rdMulti = 100.0 / (cbBuf ? cbBuf : 1);
+    double rdMulti = 100.0 / (cbBuf ? (double)cbBuf : 1.0);
 
     /* Working buffer */
     char *pvTmp = (char*)pvBuf;
@@ -74,7 +84,7 @@ RTR3DECL(int) RTSha1Digest(void* pvBuf, size_t cbBuf, char **ppszDigest, PFNRTPR
         /* Call the progress callback if one is defined */
         if (pfnProgressCallback)
         {
-            rc = pfnProgressCallback((unsigned)(cbReadTotal * rdMulti), pvUser);
+            rc = pfnProgressCallback((unsigned)((double)cbReadTotal * rdMulti), pvUser);
             if (RT_FAILURE(rc))
                 break; /* canceled */
         }
@@ -119,7 +129,7 @@ RTR3DECL(int) RTSha1DigestFromFile(const char *pszFile, char **ppszDigest, PFNRT
         return rc;
 
     /* Fetch the file size. Only needed if there is a progress callback. */
-    double rdMulti = 0;
+    double rdMulti = 0.0;
     if (pfnProgressCallback)
     {
         uint64_t cbFile;
@@ -129,7 +139,7 @@ RTR3DECL(int) RTSha1DigestFromFile(const char *pszFile, char **ppszDigest, PFNRT
             RTFileClose(hFile);
             return rc;
         }
-        rdMulti = 100.0 / (cbFile ? cbFile : 1);
+        rdMulti = 100.0 / (cbFile ? (double)cbFile : 1.0);
     }
 
     /* Allocate a reasonably large buffer, fall back on a tiny one. */
@@ -160,7 +170,7 @@ RTR3DECL(int) RTSha1DigestFromFile(const char *pszFile, char **ppszDigest, PFNRT
         /* Call the progress callback if one is defined */
         if (pfnProgressCallback)
         {
-            rc = pfnProgressCallback((unsigned)(cbReadTotal * rdMulti), pvUser);
+            rc = pfnProgressCallback((unsigned)((double)cbReadTotal * rdMulti), pvUser);
             if (RT_FAILURE(rc))
                 break; /* canceled */
         }
